Deletion nested set update disabling, refs #10962
Added the suppport, to models generated by the QubitObjectBuilder class, to
disable nested set updating during deletion. Updated existng model code.

+++ lib/model/om/BaseInformationObject.php
@@ -461,7 +461,10 @@ abstract class BaseInformationObject extends QubitObject implements ArrayAccess
     }
 
     $this->clear();
-    $this->deleteFromNestedSet($connection);
+    if (!property_exists($this, 'disableNestedSetUpdating') || $this->disableNestedSetUpdating !== true)
+    {
+      $this->deleteFromNestedSet($connection);
+    }
 
     parent::delete($connection);
