+++ lib/model/om/BaseTaxonomy.php
@@ -421,7 +421,10 @@ abstract class BaseTaxonomy extends QubitObject implements ArrayAccess
     }
 
     $this->clear();
-    $this->deleteFromNestedSet($connection);
+    if (!property_exists($this, 'disableNestedSetUpdating') || $this->disableNestedSetUpdating !== true)
+    {
+      $this->deleteFromNestedSet($connection);
+    }
 
     parent::delete($connection);
