+++ lib/propel/builder/QubitObjectBuilder.php
@@ -1698,7 +1698,10 @@ script;
       $script .= <<<script
 
     \$this->clear();
-    \$this->deleteFromNestedSet(\$connection);
+    if (!property_exists(\$this, 'disableNestedSetUpdating') || \$this->disableNestedSetUpdating !== true)
+    {
+      \$this->deleteFromNestedSet(\$connection);
+    }
 
 script;
     }
