+++ plugins/qbAclPlugin/lib/model/om/BaseAclGroup.php
@@ -761,7 +761,10 @@ abstract class BaseAclGroup implements ArrayAccess
     }
 
     $this->clear();
-    $this->deleteFromNestedSet($connection);
+    if (!property_exists($this, 'disableNestedSetUpdating') || $this->disableNestedSetUpdating !== true)
+    {
+      $this->deleteFromNestedSet($connection);
+    }
 
     $criteria = new Criteria;
     $criteria->add(QubitAclGroup::ID, $this->id);
